+++ app/(home)/(routes)/countries/layout.tsx
@@ -2,7 +2,7 @@
 
 import { useMediaQuery } from "@react-hook/media-query";
 
-import { sortedAllCountries } from "./[code]/page";
+import { sortedAllCountries } from "./_data/dummy";
 import { MobileMenu } from "@/components/mobile-menu";
 import { Menu } from "@/components/menu";
